Revert using plaid transaction id for ynab integration id
Apparently plaid's transaction id can change for pending transactions so it's not
a reliable identifier. This reverts back to using values from the transaction instead.
Previously it included the date in the identifier but sometimes the date on the plaid
transaction changes so now this only uses amount and description.