Set "multiplier" in ynab account note
I added a mortgage account to be tracked by balance, but plaid shows its
balance as positive even though it's a debt. My simple solution was to set
a multiplier in the ynab account note after the account name. If no multiplier
is found then it defaults to 1.

Example note: "Bank ABC|-1"
^ Will multiply the balance from plaid by -1